Fix bug and add more tests
